+++ containerscan/gojayunmarshaller.go
@@ -241,21 +241,3 @@ func (scan *ScanResultReport) UnmarshalJSONObject(dec *gojay.Decoder, key string
 func (scan *ScanResultReport) NKeys() int {
 	return 0
 }
-
-func (scan *ScanResultReportV1) UnmarshalJSONObject(dec *gojay.Decoder, key string) (err error) {
-	switch key {
-	case "timestamp":
-		err = dec.Int64(&(scan.Timestamp))
-	case "containersScanID":
-		err = dec.String(&(scan.ContainerScanID))
-	case "designators":
-		err = dec.Object(&(scan.Designators))
-	case "hasRelevancyData":
-		err = dec.Bool(&(scan.HasRelevancyData))
-	}
-	return err
-}
-
-func (scan *ScanResultReportV1) NKeys() int {
-	return 3
-}
