+++ containerscan/interfaces.go
